requirements.txt
Flask==1.0.2
redis==3.1.0
requests==2.21.0
pyHS100==0.3.4
PyJWT==1.7.1
